Convert cimg class into grayscale xy matrix. Helper function for locate_image(). Use grayscale to Speed up and to simplify code.

Convert cimg class into grayscale xy matrix. Helper function for locate_image(). Use grayscale to Speed up and to simplify code.

Usage

image2gray_matrix(img)

Arguments

img

A cimg object.

Value

 An xy dimensional matrix.